Move default middleware
